+++ src/query/service/tests/it/api/http/cluster.rs
@@ -16,7 +16,7 @@ use common_base::base::tokio;
 use common_exception::Result;
 use common_meta_types::NodeInfo;
 use databend_query::api::http::v1::cluster::*;
-use databend_query::test_utils::TestGlobalServices;
+use databend_query::test_kits::TestGlobalServices;
 use poem::get;
 use poem::http::header;
 use poem::http::Method;
@@ -30,7 +30,7 @@ use pretty_assertions::assert_eq;
 #[tokio::test(flavor = "multi_thread")]
 async fn test_cluster() -> Result<()> {
     let _guard =
-        TestGlobalServices::setup(databend_query::test_utils::ConfigBuilder::create().build())
+        TestGlobalServices::setup(databend_query::test_kits::ConfigBuilder::create().build())
             .await?;
     let cluster_router = Route::new().at("/v1/cluster/list", get(cluster_list_handler));
