+++ src/query/service/tests/it/api/http/logs.rs
@@ -15,7 +15,7 @@
 use common_base::base::tokio;
 use common_exception::Result;
 use databend_query::api::http::v1::logs::logs_handler;
-use databend_query::test_utils::TestGlobalServices;
+use databend_query::test_kits::TestGlobalServices;
 use poem::get;
 use poem::http::Method;
 use poem::http::StatusCode;
@@ -28,7 +28,7 @@ use pretty_assertions::assert_eq;
 #[tokio::test(flavor = "multi_thread")]
 async fn test_logs() -> Result<()> {
     let _guard =
-        TestGlobalServices::setup(databend_query::test_utils::ConfigBuilder::create().build())
+        TestGlobalServices::setup(databend_query::test_kits::ConfigBuilder::create().build())
             .await?;
 
     let test_router = Route::new().at("/v1/logs", get(logs_handler));
