+++ src/query/service/tests/it/servers/flight_sql/flight_sql_handler.rs
@@ -29,8 +29,8 @@ use common_exception::Result;
 use common_meta_app::principal::AuthInfo;
 use common_meta_app::principal::PasswordHashMethod;
 use databend_query::servers::flight_sql::flight_sql_service::FlightSqlServiceImpl;
-use databend_query::test_utils::ConfigBuilder;
-use databend_query::test_utils::TestGlobalServices;
+use databend_query::test_kits::ConfigBuilder;
+use databend_query::test_kits::TestGlobalServices;
 use futures::TryStreamExt;
 use goldenfile::Mint;
 use tempfile::NamedTempFile;
