+++ src/query/service/tests/it/servers/http/clickhouse_handler.rs
@@ -21,8 +21,8 @@ use databend_query::servers::http::middleware::HTTPSessionMiddleware;
 use databend_query::servers::http::v1::clickhouse_router;
 use databend_query::servers::http::CLICKHOUSE_VERSION;
 use databend_query::servers::HttpHandlerKind;
-use databend_query::test_utils::ConfigBuilder;
-use databend_query::test_utils::TestGlobalServices;
+use databend_query::test_kits::ConfigBuilder;
+use databend_query::test_kits::TestGlobalServices;
 use http::Uri;
 use poem::error::Result as PoemResult;
 use poem::http::Method;
