+++ src/query/service/tests/it/servers/http/http_query_handlers.rs
@@ -40,8 +40,8 @@ use databend_query::servers::http::v1::QueryResponse;
 use databend_query::servers::HttpHandler;
 use databend_query::servers::HttpHandlerKind;
 use databend_query::sessions::QueryAffect;
-use databend_query::test_utils::ConfigBuilder;
-use databend_query::test_utils::TestGlobalServices;
+use databend_query::test_kits::ConfigBuilder;
+use databend_query::test_kits::TestGlobalServices;
 use headers::Header;
 use jwt_simple::algorithms::RS256KeyPair;
 use jwt_simple::algorithms::RSAKeyPairLike;
