+++ src/query/service/tests/it/servers/mysql/mysql_handler.rs
@@ -23,8 +23,8 @@ use common_exception::ErrorCode;
 use common_exception::Result;
 use common_exception::ToErrorCode;
 use databend_query::servers::MySQLHandler;
-use databend_query::test_utils::ConfigBuilder;
-use databend_query::test_utils::TestGlobalServices;
+use databend_query::test_kits::ConfigBuilder;
+use databend_query::test_kits::TestGlobalServices;
 use mysql_async::prelude::FromRow;
 use mysql_async::prelude::Queryable;
 use mysql_async::FromRowError;
