+++ src/query/service/tests/it/sessions/query_ctx.rs
@@ -33,7 +33,7 @@ async fn test_get_storage_accessor_s3() -> Result<()> {
         .mount(&mock_server)
         .await;
 
-    let mut conf = databend_query::test_utils::ConfigBuilder::create().config();
+    let mut conf = databend_query::test_kits::ConfigBuilder::create().config();
 
     conf.storage.params = StorageParams::S3(StorageS3Config {
         region: "us-east-2".to_string(),
@@ -44,7 +44,7 @@ async fn test_get_storage_accessor_s3() -> Result<()> {
     });
 
     let (_guard, qctx) =
-        databend_query::test_utils::create_query_context_with_config(conf, None).await?;
+        databend_query::test_kits::create_query_context_with_config(conf, None).await?;
 
     let _ = qctx.get_data_operator()?;
 
@@ -53,14 +53,14 @@ async fn test_get_storage_accessor_s3() -> Result<()> {
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 async fn test_get_storage_accessor_fs() -> Result<()> {
-    let mut conf = databend_query::test_utils::ConfigBuilder::create().config();
+    let mut conf = databend_query::test_kits::ConfigBuilder::create().config();
 
     conf.storage.params = StorageParams::Fs(StorageFsConfig {
         root: "/tmp".to_string(),
     });
 
     let (_guard, qctx) =
-        databend_query::test_utils::create_query_context_with_config(conf, None).await?;
+        databend_query::test_kits::create_query_context_with_config(conf, None).await?;
 
     let _ = qctx.get_data_operator()?;
