+++ src/query/service/tests/it/storages/fuse/operations/alter_table.rs
@@ -38,7 +38,7 @@ use databend_query::interpreters::AddTableColumnInterpreter;
 use databend_query::interpreters::DropTableColumnInterpreter;
 use databend_query::interpreters::Interpreter;
 use databend_query::interpreters::InterpreterFactory;
-use databend_query::test_utils::table_test_fixture::TestFixture;
+use databend_query::test_kits::table_test_fixture::TestFixture;
 use futures_util::TryStreamExt;
 use ordered_float::OrderedFloat;
 use storages_common_cache::LoadParams;
