+++ src/query/service/tests/it/storages/fuse/operations/commit.rs
@@ -83,8 +83,8 @@ use common_storages_fuse::FuseTable;
 use common_storages_fuse::FUSE_TBL_SNAPSHOT_PREFIX;
 use dashmap::DashMap;
 use databend_query::sessions::QueryContext;
-use databend_query::test_utils::table_test_fixture::execute_query;
-use databend_query::test_utils::table_test_fixture::TestFixture;
+use databend_query::test_kits::table_test_fixture::execute_query;
+use databend_query::test_kits::table_test_fixture::TestFixture;
 use futures::TryStreamExt;
 use rand::thread_rng;
 use rand::Rng;
