+++ src/query/service/tests/it/storages/fuse/operations/gc.rs
@@ -25,9 +25,9 @@ use common_storages_fuse::io::SegmentWriter;
 use common_storages_fuse::statistics::gen_columns_statistics;
 use common_storages_fuse::statistics::merge_statistics;
 use common_storages_fuse::FuseTable;
-use databend_query::test_utils::table_test_fixture::append_sample_data;
-use databend_query::test_utils::table_test_fixture::check_data_dir;
-use databend_query::test_utils::table_test_fixture::TestFixture;
+use databend_query::test_kits::table_test_fixture::append_sample_data;
+use databend_query::test_kits::table_test_fixture::check_data_dir;
+use databend_query::test_kits::table_test_fixture::TestFixture;
 use futures_util::TryStreamExt;
 use storages_common_table_meta::meta::Location;
 use storages_common_table_meta::meta::SegmentInfo;
