+++ src/query/service/tests/it/storages/fuse/operations/purge_truncate.rs
@@ -13,11 +13,11 @@
 //  limitations under the License.
 use common_base::base::tokio;
 use common_exception::Result;
-use databend_query::test_utils::table_test_fixture::append_sample_data;
-use databend_query::test_utils::table_test_fixture::check_data_dir;
-use databend_query::test_utils::table_test_fixture::execute_command;
-use databend_query::test_utils::table_test_fixture::history_should_have_item;
-use databend_query::test_utils::table_test_fixture::TestFixture;
+use databend_query::test_kits::table_test_fixture::append_sample_data;
+use databend_query::test_kits::table_test_fixture::check_data_dir;
+use databend_query::test_kits::table_test_fixture::execute_command;
+use databend_query::test_kits::table_test_fixture::history_should_have_item;
+use databend_query::test_kits::table_test_fixture::TestFixture;
 
 #[tokio::test(flavor = "multi_thread")]
 async fn test_fuse_truncate_purge_stmt() -> Result<()> {
