+++ src/query/service/tests/it/storages/fuse/operations/table_analyze.rs
@@ -26,9 +26,9 @@ use databend_query::sessions::QueryContext;
 use databend_query::sessions::TableContext;
 use databend_query::sql::plans::Plan;
 use databend_query::sql::Planner;
-use databend_query::test_utils::table_test_fixture::execute_command;
-use databend_query::test_utils::table_test_fixture::execute_query;
-use databend_query::test_utils::table_test_fixture::TestFixture;
+use databend_query::test_kits::table_test_fixture::execute_command;
+use databend_query::test_kits::table_test_fixture::execute_query;
+use databend_query::test_kits::table_test_fixture::TestFixture;
 use futures_util::TryStreamExt;
 
 use crate::storages::fuse::operations::mutation::do_deletion;
