+++ src/query/service/tests/it/storages/fuse/operations/truncate.rs
@@ -19,8 +19,8 @@ use common_catalog::table::Table;
 use common_exception::ErrorCode;
 use common_expression::DataBlock;
 use common_sql::executor::table_read_plan::ToReadDataSourcePlan;
-use databend_query::test_utils::table_test_fixture::execute_query;
-use databend_query::test_utils::table_test_fixture::TestFixture;
+use databend_query::test_kits::table_test_fixture::execute_query;
+use databend_query::test_kits::table_test_fixture::TestFixture;
 use futures_util::TryStreamExt;
 
 #[tokio::test(flavor = "multi_thread")]
