+++ src/query/service/tests/it/table_functions/numbers_table.rs
@@ -29,7 +29,7 @@ use pretty_assertions::assert_eq;
 #[tokio::test(flavor = "multi_thread")]
 async fn test_number_table() -> Result<()> {
     let tbl_args = TableArgs::new_positioned(vec![Scalar::from(8u64)]);
-    let (_guard, ctx) = databend_query::test_utils::create_query_context().await?;
+    let (_guard, ctx) = databend_query::test_kits::create_query_context().await?;
     let table = NumbersTable::create("system", "numbers_mt", 1, tbl_args)?;
 
     let source_plan = table
